[Dev] mail

+++ src/main/java/me/qfdk/payment/controller/PageController.java
@@ -1,6 +1,8 @@
 package me.qfdk.payment.controller;
 
+import me.qfdk.payment.entity.MyPayment;
 import me.qfdk.payment.entity.Product;
+import me.qfdk.payment.repository.PaymentRepository;
 import me.qfdk.payment.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,12 +12,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
 import java.math.BigDecimal;
+import java.util.List;
 
 @Controller
 public class PageController {
 
     @Autowired
     private ProductRepository productRepository;
+
+    @Autowired
+    private PaymentRepository paymentRepository;
 
     @Value("${adminToken}")
     private String myToken;
@@ -37,7 +43,16 @@ public class PageController {
             String id = req.getParameter("id");
 
             if ("delete".equals(action)) {
-                productRepository.deleteById(Integer.parseInt(id));
+                List<MyPayment> list = paymentRepository.findByProductId(Integer.parseInt(id));
+                if (list.isEmpty()) {
+                    if (productRepository.findById(Integer.parseInt(id)).isPresent()) {
+                        productRepository.deleteById(Integer.parseInt(id));
+                    }
+                } else {
+                    model.addAttribute("products", productRepository.findAll());
+                    model.addAttribute("error", "支付记录已存在该产品，请手动删除该记录");
+                    return "admin";
+                }
             }
 
             if ("add".equals(action)) {
